+++ Cargo.toml
@@ -1,6 +1,6 @@
 [workspace]
 resolver = "2"
-members = [ "tigtag3d", "public", "macros", "macros_inside" ]
+members = [ "tigtag3d", "tigtag_inside", "public", "macros", "macros_inside" ]
 
 [workspace.package]
 version = "0.1.0"
@@ -18,6 +18,7 @@ syn = "2"
 quote = "1"
 proc-macro2 = "1"
 
+tigtag_inside = { path = "tigtag_inside" }
 public        = { path = "public"        }
 macros        = { path = "macros"        }
 macros_inside = { path = "macros_inside" }
