+++ public/src/config.rs
@@ -6,6 +6,11 @@ use super::*;
 pub const APP_TITLE: &str = CARGO_TOML_NAME; //アプリタイトル
 pub const APP_VER  : &str = CARGO_TOML_VER;  //アプリのバージョン
 
+//タイトルロゴ
+pub const GAME_TITLE_LOGO: &str = "tigtag3D";
+
+////////////////////////////////////////////////////////////////////////////////
+
 //単位Gridの縦横(Pixel)
 const BASE_PIXELS: i32 = 8;
 const SCALING: f32 = 4.0;
@@ -47,6 +52,69 @@ pub const GRIDS_Y_RANGE: Range<i32> = 0..SCREEN_GRIDS_HEIGHT;
 //ログレベル
 pub const LOG_LEVEL_DEV: &str = "warn,wgpu_hal=error"; //開発
 pub const LOG_LEVEL_REL: &str = "error"; //リリース
+
+////////////////////////////////////////////////////////////////////////////////
+
+//画面デザイン(枠)
+pub const SCREEN_FRAME_SPACE_CHAR : char = ' ';
+pub const SCREEN_FRAME_LABEL_REGEX: &str = r"[a-zA-Z0-9\.]+";
+
+pub const VIEWPORT_MINIMAP_ORIGIN: IVec2 = IVec2::new( 33, 13 );
+pub const VIEWPORT_MINIMAP_SIZE  : IVec2 = IVec2::new(  9,  9 );
+
+pub static SCREEN_FRAME: Lazy<ScreenFrame> = Lazy::new
+(   ||
+    {   let design = vec!
+        [  //0123456789_123456789_123456789_123456789_12
+            "###########################################", //0
+            "#                               ###########", //1
+            "#                               ###########", //2
+            "#                               ###########", //3
+            "#                               ###########", //4
+            "#                               ###########", //5
+            "#                               ###########", //6
+            "#                               ###########", //7
+            "#                               ###########", //8
+            "#                               ###########", //9
+            "#                               ###########", //10
+            "#                               ###########", //11
+            "#                               #tigtag2D##", //12
+            "#                               #         #", //13
+            "#                               #         #", //14
+            "#                               #         #", //15
+            "#                               #         #", //16
+            "#                               #         #", //17
+            "#                               #         #", //18
+            "#                               #         #", //19
+            "#                               #         #", //20
+            "#                               #         #", //21
+            "###########################################", //22
+            "                                           ", //23
+        ]; //0123456789_123456789_123456789_123456789_12
+
+        if design[ 0 ].len() != SCREEN_GRIDS_WIDTH  as usize
+        || design.len()      != SCREEN_GRIDS_HEIGHT as usize
+        {   panic!( "APPERR: {}", ER_BAD_SCREEN_DESIGN );
+        }
+
+        //3Dカメラの表示領域(viewport)の設定
+        let viewport = ViewPortInfo
+        {   origin: ( IVec2::new(  1,  1 ).as_vec2() - 0.5 ) * PIXELS_PER_GRID,
+            size  : ( IVec2::new( 31, 21 ).as_vec2() + 1.0 ) * PIXELS_PER_GRID,
+        };
+
+        //ミニマップの小窓の設定
+        let minimap = MiniMapInfo
+        {   zero: VIEWPORT_MINIMAP_ORIGIN,
+            size: VIEWPORT_MINIMAP_SIZE,
+        };
+
+        ScreenFrame { design, viewport, minimap }
+    }
+);
+
+//エラーメッセージ
+const ER_BAD_SCREEN_DESIGN: &str = "Frame design unmatch width/height parameters.";
 
 ////////////////////////////////////////////////////////////////////////////////
