+++ tigtag3d/src/init_app.rs
@@ -61,24 +61,6 @@ impl Plugin for Schedule
 
 ////////////////////////////////////////////////////////////////////////////////
 
-//InitAppの後の遷移先を登録するResource
-#[derive( Resource )]
-pub struct AfterInitApp ( pub MyState );
-impl ChangeMyState for AfterInitApp
-{   fn state( &self ) -> MyState { self.0 }
-}
-
-//隠しノードのComponent
-#[derive( Component )] pub struct HiddenNode;
-
-//ヘッダー／フッターのComponent
-#[derive( Component )] pub struct UiHeaderLeft;   //日時表示
-#[derive( Component )] pub struct UiHeaderCenter; //タイトル
-#[derive( Component )] pub struct UiHeaderRight;  //経過時間表示
-#[derive( Component )] pub struct UiFooterLeft;   //FPS表示
-#[derive( Component )] pub struct UiFooterCenter; //auther
-#[derive( Component )] pub struct UiFooterRight;  //Powered by
-
 //ヘッダー／フッターのテキスト情報
 const HEADER_LEFT: &[ MessageSect ] =
 &[  ( "  Clock ", ASSETS_FONT_ORBITRON_BLACK      , PIXELS_PER_GRID * 0.6, Color::TEAL   ),
