+++ tigtag3d/src/main.rs
@@ -6,38 +6,25 @@ use bevy::
     asset::{ LoadState, LoadedUntypedAsset },
     diagnostic::{ FrameTimeDiagnosticsPlugin, DiagnosticsStore },
     utils::Duration,
-  //=============================================================
-    sprite::MaterialMesh2dBundle,
     utils::{ HashMap, HashSet },
-    audio::Volume,
-    input::keyboard::NativeKeyCode,
-  //=============================================================
     sprite::Anchor,
     render::camera::Viewport,
 };
 use once_cell::sync::Lazy;
 use rand::prelude::*;
-//========================
 use regex::Regex;
 
 //standard library
-use std::
-{   ops::Range,
-    f32::consts::{ PI, TAU },
-  //=========================
-    ops::{ Add, AddAssign },
-    cmp::Ordering,
-    collections::VecDeque,
-};
+use std::f32::consts::TAU;
 
 //import names from other crates in this package
 use public::*;
+use tigtag_inside as tigtag;
 
 //internal submodules
 mod debug;
 mod load_assets;
 mod init_app;
-mod tigtag;
 mod tigtag3d;
 
 ////////////////////////////////////////////////////////////////////////////////
