--- tigtag3d/src/tigtag/mod.rs
@@ -1,32 +1,0 @@
-use super::*;
-
-//ゲームロジック
-mod schedule;
-pub use schedule::Schedule;
-
-//play_game内共通
-mod common;
-use common::*;
-
-//UIの処理
-mod header;
-mod ui;
-use ui::*;
-
-//pause処理
-mod pause;
-
-//マップ、自機、追手の処理
-mod map;
-use map::GridToPixelOnMap;
-mod player;
-mod chasers;
-mod detection;
-
-//デモ
-mod demo;
-
-//名前の輸出用prelude
-pub mod prelude;
-
-//End of code.
