+++ tigtag_inside/src/pause.rs
@@ -123,7 +123,7 @@ impl effect::Scaling for SelectedMenuItem
 
 //PAUSEメニューをspawnする
 fn spawn_ui_text
-(   qry_hidden_node: Query<Entity, With<init_app::HiddenNode>>,
+(   qry_hidden_node: Query<Entity, With<HiddenNode>>,
     mut cmds: Commands,
     asset_svr: Res<AssetServer>,
 )
