+++ tigtag_inside/src/ui/game_over.rs
@@ -75,7 +75,7 @@ impl effect::Blinking for TextREPLAY
 
 //ゲームオーバーをspawnする
 pub fn spawn_text
-(   qry_hidden_node: Query<Entity, With<init_app::HiddenNode>>,
+(   qry_hidden_node: Query<Entity, With<HiddenNode>>,
     mut cmds: Commands,
     asset_svr: Res<AssetServer>,
 )
