+++ tigtag_inside/src/ui/stage_start.rs
@@ -51,7 +51,7 @@ impl<'a> effect::CountDown for CountDown<'a>
 
 //ステージスタートをspawnする
 pub fn spawn_text
-(   qry_hidden_node: Query<Entity, With<init_app::HiddenNode>>,
+(   qry_hidden_node: Query<Entity, With<HiddenNode>>,
     mut cmds: Commands,
     asset_svr: Res<AssetServer>,
 )
