+++ tigtag_inside/src/ui/title_demo.rs
@@ -53,7 +53,7 @@ impl effect::Blinking for TextDEMO
 
 //タイトルをspawnする
 pub fn spawn_text
-(   qry_hidden_node: Query<Entity, With<init_app::HiddenNode>>,
+(   qry_hidden_node: Query<Entity, With<HiddenNode>>,
     mut cmds: Commands,
     asset_svr: Res<AssetServer>,
 )
